fix: set mode bit before changing duty cycle
When the device is set to RPM control mode, the driver does not
cleanly apply the new duty cycle when changing back to duty cycle
control.

The IC appears to set the fan drive register to the last duty cycle
that was being used when targeting an RPM. Change the mode bit first
before setting a new duty cycle.